### 2018-01-28 #553 @whedon
[PRE REVIEW]: PetIBM: toolbox and applications of the immersed-boundary method on distributed-memory architectures
**Submitting author:** @labarba (<a href="http://orcid.org/0000-0001-5812-2711">Lorena A. Barba</a>)
**Repository:** <a href="https://github.com/barbagroup/PetIBM" target ="_blank">https://github.com/barbagroup/PetIBM</a>
**Version:** v0.3
