### 2018-01-28 #553 @labarba
We suggested @kyleniemeyer as editor, since he edited our previous submission of a related code:
https://github.com/openjournals/joss-reviews/issues/301

For reviewer, we think @jedbrown would be great, if he's willing. He reviewed
https://github.com/openjournals/joss-reviews/issues/280
... which is used in PetIBM.
