### 2018-01-29 #553 @kyleniemeyer
Sure, I can review this one.
