### 2018-01-29 #553 @labarba
Just wanted to cancel my comment to say that I suggested you as editor, rather than reviewer.

We really need more editors for computational mechanics and HPC.
