### 2018-01-29 #558 @whedon
[REVIEW]: PetIBM: toolbox and applications of the immersed-boundary method on distributed-memory architectures
**Submitting author:** @labarba (<a href="http://orcid.org/0000-0001-5812-2711">Lorena A. Barba</a>)
**Repository:** <a href="https://github.com/barbagroup/PetIBM" target ="_blank">https://github.com/barbagroup/PetIBM</a>
**Version:** v0.3
