### 2018-01-29 #553 @kyleniemeyer
ok @labarba @jedbrown we'll proceed with the actual review in #558
