### 2018-03-13 b35d274
fix memory leak coming from CartesianMesh and in NavierStokesSolver
Fix memory leaks coming from Mesh/CartesianMesh. This memory leak didn't
affect simulations. It just didn't free all memory before the end of the
porgrams. And those memory would be freed automatically after the
program is killed.

- fix: `dim` was set to -1 before being used in the loop that destroys
  the underlying DMDAs for velocity fileds.
- Manually launch destroying process: We need to destroy PETSc objects
  before calling PetscFinalize(). And calling deconstructor manually is
  not encouraged by C++ communicties, so w
[…]
