### 2018-03-13 49528a4
fix memory leak in IBPM solver
This memory leak didn't affect simulation results.

- manually destroy PETSc objects
- correctly release the pressure vector created with DMDA
