### 2018-03-14 460a209
delete automatic destroying of AmgXSolver and fix a typo in config.yaml
