### 2018-03-19 #558 @labarba
Hi @jedbrown -- @mesnardo and @piyueh have been working on the `joss-review` branch, addressing the issues you opened on the code repo itself.
