### 2018-04-23 3530698
Remove dependence on env var PETIBM_EXAMPLES
* Remove the dependence on the environment variable `PETIBM_EXAMPLES`
in the Python script.
* Use `os.path.dirname()` wherever appropriate in the Python scripts.
* Update the documentation and add more explanation about the outputs
of the PetIBM post-processing utilities `petibm-vorticity` and
`petibm-createxdmf`.
