+++ examples/decoupledibpm/cylinder2dRe550_GPU/scripts/createBody.py
@@ -12,7 +12,7 @@ xc, yc = 0.0, 0.0  # center's coordinates
 ds = 0.01  # distance between two consecutive points
 
 script_dir = os.path.dirname(os.path.realpath(__file__))
-simu_dir = os.sep.join(script_dir.split(os.sep)[:-1])
+simu_dir = os.path.dirname(script_dir)
 
 # Generate coordinates of the circle.
 n = int(2.0 * numpy.pi * R / ds)
