### 2018-04-23 c414d6c
make the XDMF works for both ParaView and VisIt
With 2DRectMesh and VXVY, ParaView and VisIt have different reading behaviors
regarding the order of axes. But with 3DRectMesh and VXVYVZ, there's no
such problem. So we now use 3DRectMesh and VXVYVZ for everything, even
it's a 2D simulation. For 2D simulations, a dummy z-axis grid line is
added with only one grid point, which has 0.0 coordinate.
