+++ applications/createxdmf/main.cpp
@@ -21,48 +21,48 @@
 /**
  * \defgroup createxdmf Post-processing utility: createxdmf
  * \brief A post-processing utility that creates XDMF files for visualization.
- * 
- * This is a helper utility built with PetIBM components, and it creates XDMF 
+ *
+ * This is a helper utility built with PetIBM components, and it creates XDMF
  * files based on the HDF5 files produced by
  * \ref nssolver "Navier-Stokes solver",
- * \ref tairacolonius "IBPM solver", and 
+ * \ref tairacolonius "IBPM solver", and
  * \ref decoupledibpm "decoupled IBPM solver".
- * XDMF files can be used for visualization in 
+ * XDMF files can be used for visualization in
  * [VisIt](https://wci.llnl.gov/simulation/computer-codes/visit/)
- * 
+ *
  * If readers are interested in using this utility,
- * please refer to 
+ * please refer to
  * \ref md_doc_markdowns_runpetibm "Running PetIBM",
  * \ref md_doc_markdowns_examples2d "2D Exmaples", and
  * \ref md_doc_markdowns_examples3d "3D Examples".
- * 
+ *
  * \ingroup apps
  */
 
 PetscErrorCode writeSingleXDMF(
-        const std::string &directory, const std::string &name, 
+        const std::string &directory, const std::string &name,
         const PetscInt &dim, const petibm::type::IntVec1D &n,
         const PetscInt &bg, const PetscInt &ed, const PetscInt &step);
 
 int main(int argc, char **argv)
 {
-    
+
     PetscErrorCode      ierr;
-    
+
     ierr = PetscInitialize(&argc, &argv, nullptr, nullptr); CHKERRQ(ierr);
-    
-    
-    
+
+
+
     YAML::Node  setting;
-    
+
     ierr = petibm::parser::getSettings(setting); CHKERRQ(ierr);
-    
+
     petibm::type::Mesh  mesh;
-    
+
     ierr = petibm::mesh::createMesh(PETSC_COMM_WORLD, setting, mesh); CHKERRQ(ierr);
-    
-    
-    
+
+
+
     PetscInt                    bg;
     PetscInt                    ed;
     PetscInt                    step;
@@ -73,43 +73,43 @@ int main(int argc, char **argv)
     if (! isSet) bg = setting["parameters"]["startStep"].as<PetscInt>(0);
 
     ierr = PetscOptionsGetInt(nullptr, nullptr, "-ed", &ed, &isSet); CHKERRQ(ierr);
-    if (! isSet) ed = bg + setting["parameters"]["nt"].as<PetscInt>(); 
+    if (! isSet) ed = bg + setting["parameters"]["nt"].as<PetscInt>();
 
     ierr = PetscOptionsGetInt(nullptr, nullptr, "-step", &step, &isSet); CHKERRQ(ierr);
-    if (! isSet) step = setting["parameters"]["nsave"].as<PetscInt>(); 
-    
-    
-    
+    if (! isSet) step = setting["parameters"]["nsave"].as<PetscInt>();
+
+
+
     // u
     ierr = writeSingleXDMF(setting["directory"].as<std::string>(),
             "u", mesh->dim, mesh->n[0], bg, ed, step); CHKERRQ(ierr);
-    
+
     // v
     ierr = writeSingleXDMF(setting["directory"].as<std::string>(),
[…]
