### 2018-04-23 6f994c2
add information about post processing with ParaView
