### 2018-04-24 c6fcfef
make XDMF files follow DTD file
- work with ParaView XDMF3 reader. Previously it only work with XDMF2
  reader.
- replace spaces with tabs.
- use entity instead of reference to avoid repeating topologies and
  geometries. This avoid xmllint's complaing of undeclare attributes
  because the DTD file provided by Kitware is not complete.
- modify ParaView information to documentation
