+++ applications/createxdmf/main.cpp
@@ -151,104 +151,89 @@ PetscErrorCode writeSingleXDMF(
     // write macro definitions
     ierr = PetscViewerASCIIPrintf(viewer,
             "<!DOCTYPE Xdmf SYSTEM \"Xdmf.dtd\" [\n"); CHKERRQ(ierr);
-    ierr = PetscViewerASCIIPrintf(viewer,
-            "    <!ENTITY CaseDir \"%s\">\n", "./"); CHKERRQ(ierr);
+    ierr = PetscViewerASCIIPrintf(viewer, "\t"
+            "<!ENTITY CaseDir \"%s\">\n", "./"); CHKERRQ(ierr);
 
     // always use 3D XDMF format, so both Visit and Paraview works
     for(int i=0; i<3; ++i)
     {
-        ierr = PetscViewerASCIIPrintf(viewer, "    <!ENTITY N%s \"%d\">\n",
+        ierr = PetscViewerASCIIPrintf(viewer, "\t"
+                "<!ENTITY N%s \"%d\">\n",
                 type::dir2str[type::Dir(i)].c_str(), n[i]); CHKERRQ(ierr);
     }
-    ierr = PetscViewerASCIIPrintf(viewer, "]>\n\n"); CHKERRQ(ierr);
 
-    // write Xdmf block
-    ierr = PetscViewerASCIIPrintf(viewer,
-            "<Xdmf Version=\"2.2\">\n"); CHKERRQ(ierr);
-    ierr = PetscViewerASCIIPrintf(viewer, "    "
-            "<Information Name=\"MeteData\" Value=\"ID-23454\"/>\n"); CHKERRQ(ierr);
-    ierr = PetscViewerASCIIPrintf(viewer, "    " "<Domain>\n\n"); CHKERRQ(ierr);
+    // topology
+    ierr = PetscViewerASCIIPrintf(viewer, "\t"
+            "<!ENTITY Topo \"<Topology TopologyType=\'3DRectMesh\' "
+            "Dimensions=\'&Nz; &Ny; &Nx;\'/>\">\n"); CHKERRQ(ierr);
 
-    // write topology; always use 3DRectMesh
-    ierr = PetscViewerASCIIPrintf(viewer, "    "
-            "<Topology Name=\"%s Topo\" ", name.c_str()); CHKERRQ(ierr);
-    ierr = PetscViewerASCIIPrintf(viewer,
-            "TopologyType=\"3DRectMesh\" "); CHKERRQ(ierr);
-    ierr = PetscViewerASCIIPrintf(viewer,
-            "NumberOfElements=\"&Nz; &Ny; &Nx;\"/>\n\n"); CHKERRQ(ierr);
+    // geometry
+    ierr = PetscViewerASCIIPrintf(viewer, "\t<!ENTITY Geo\n"); CHKERRQ(ierr);
+    ierr = PetscViewerASCIIPrintf(viewer, "\t\t"
+            "\"<Geometry GeometryType=\'VXVYVZ\'>\n"); CHKERRQ(ierr);
 
-    // write geometry; always use 3D, i.e., VXVYVZ
-    ierr = PetscViewerASCIIPrintf(viewer, "    "
-            "<Geometry Name=\"%s Geo\" ", name.c_str()); CHKERRQ(ierr);
-    ierr = PetscViewerASCIIPrintf(viewer,
-            "GeometryType=\"VXVYVZ\">\n"); CHKERRQ(ierr);
-
-    // output grid coordinates
     for(int i=0; i<dim; ++i)
     {
         std::string dir = type::dir2str[type::Dir(i)];
-        ierr = PetscViewerASCIIPrintf(viewer, "    " "    "
[…]
