### 2018-05-01 #558 @jedbrown
A couple final comments after iterating in the PetIBM repository.

* I don't know why the 2D simulation is run via `petibm-tairacolonius` while the 3D uses `petibm-decoupledibpm`. Do these need to be separate executables and are these the best names for a user?

* **Authorship** The corresponding author (@labarba) has not the author or committer of any code for this project, but has advised the authors. I imagine this counts for JOSS, but it may be helpful to rephrase "made major contributions to the software?"

* It would be a lot more convenient (given vis user interfaces) if there was a single Xdmf file containing all the variables instead of separate files for each variable.

* The package installs headers and a library (shared and static, plus libtool), suggesting that calling the library (rather than running the provided executables) is an intended mode of operation. However, there is no documentation of this, instead focusing exclusively on the provided executables. If the libraries are not intended to be called by user code, I would recommend not installing the headers and only installing the used library `libpetibm.so.0`. If it is intended for use, then it should be documented and the library should be cleaned of non-namespaced symbols (e.g., `readSingleYAML`).
