### 2018-05-01 #558 @labarba
Contributions are not always commits. Discussions on design, code reviews, and project management are important contributions to software that may not reflect on git commits. JOSS recognizes this and it's up to the authors to determine.
