### 2018-05-01 #558 @jedbrown
@labarba I didn't mean to suggest otherwise, but it's hard for a reviewer to audit when looking at the repository. Some journals ask for explicit statements describing the contribution from each author. In case of contributions that are not visible in the repository (commits and citations in commit messages are the most durable record), perhaps JOSS should request such a statement. (I have no doubts about your contributions so this is purely a process matter.)
