+++ doc/markdowns/runpetibm.md
@@ -1,13 +1,32 @@
 # Run PetIBM
 
-Once PetIBM is installed, the executables should be located in the `bin` folder of your installation directory.
-For convenience, you can prepend the `bin` directory to your PATH environment variable:
+PetIBM is a library (a toolbox to solve the Navier-Stokes equations with an immersed-boundary method) that comes with several application codes.
+Upon successful installation, the library (shared and/or static) and the application programs should be respectively located in the `lib` and `bin` folders of your installation directory.
+
+Once PetIBM is installed, the libraries (shared and/or static) are located in the `lib` folder of your installation directory.
+The present software package comes with 5 application codes that use the PetIBM library.
+Upon successful installation, the binary executables for these applications are located in the `bin` folder of your installation directory.
+For convenience, you can prepend your PATH environment variable with the `bin` directory to use the binary executables:
 
     export PATH=<petibm-installation-directory>/bin:$PATH
 
-To run a pure Navier-Stokes simulation (no immersed boundary in the computational domain) in serial:
+List of binary executables:
+    * `petibm-navierstokes`
+    * `petibm-ibpm`
+    * `petibm-decoupledibpm`
+    * `petibm-vorticity`
+    * `petibm-createxdmf`
 
-## Running PetIBM
+These programs work for 2D **and** 3D configurations and can be run in serial or parallel (with `mpiexec` or `mpirun`).
+The following sub-sections provide more details about each executable.
+
+## Program `petibm-navierstokes`
+
+The program solves the 2D or 3D Navier-Stokes equations using a projection method: an approximate block-LU decomposition (Perot, 1993).
+The equations are fully discretized in space (second-order central differences) and in time.
+At each time step, the program solves a system for the velocity field and then a Poisson system for the pressure to project the velocity onto the space of divergence-free fields.
+
+To run the program:
 
     cd <simulation-directory>
     petibm-navierstokes
@@ -24,39 +43,84 @@ To run on two CPU processes:
 
     mpiexec -np 2 petibm-navierstokes
 
-To run a simulation with the Immersed Boundary Projection Method (Taira and Colonius, 2007):
 
-    mpiexec -np n petibm-tairacolonius
+## Program `petibm-ibpm`
 
-To run a simulation with a decoupled version of the Immersed Boundary Projection Method (Li et al., 2016):
+The program solves the 2D and 3D Navier-Stokes equations with an immersed-boundary method: the Immersed Boundary Projection Method (IBPM) proposed by Taira and Colonius (2007).
+The method is based on the projection approach of Perot (1993) and is suitable to solve the flow around one or several, fixed or moving, rigid immersed boundaries.
+Note that the current implementation of the method in PetIBM only works for fixed fixed immersed boundaries.
+At each time step, the program solves a system for the velocity field and then a modified Poisson system for the pressure field and Lagrangian forces.
+Finally, the velocity field  is projected onto the space of divergence-free fields that satisfy the no-slip condition at the location of the Lagrangian boundary points.
 
+To run a simulation with the IBPM:
+
+    cd <simulation-directory>
+    mpiexec -np n petibm-ibpm
+
+You can also provide the path of the simulation directory with the command-line argument `-directory <path>` and/or the path of the YAML configuration file with `-config <path>`.
+
+
+## Program `petibm-decoupledibpm`
+
[…]
