+++ examples/ibpm/cylinder2dRe550_GPU/README.md
@@ -4,7 +4,7 @@ Run the example using 2 CPU processes and 1 GPU device:
 
 ```
 export CUDA_VISIBLE_DEVICES=<idx1>
-mpiexec -np 2 petibm-tairacolonius -options_left -log_view ascii:stdout.txt
+mpiexec -np 2 petibm-ibpm -options_left -log_view ascii:stdout.txt
 ```
 
 The simulation completes in less than 5 minutes when using:
