+++ examples/api_examples/liddrivencavity2d/main.cpp
@@ -1,0 +1,382 @@
+/* A simple coding exmaple for using PetIBM API to build a Navier-Stokes solver.
+ *
+ * The numerical method used is based on the following paper:
+ *
+ * Perot, J. B. (1993). An analysis of the fractional step method. Journal of
+ * Computational Physics, 108(1), 51-58.
+ *
+ * The discretization method is central finite difference and staggered grids.
+ *
+ * \file examples/api_examples/liddrivencavity2d/main.cpp
+ * \copyright Copyright (c) 2016-2018, Barba group. All rights reserved.
+ * \license BSD 3-Clause License.
+ */
+
+
+// PetIBM relies on YAML to simulation configurations, such as mesh settings,
+// time step size, boundary conditions, etc. So we first include YAML header.
+# include <yaml-cpp/yaml.h>
+
+// End-users of a solver can provide simulation configurations through a YAML
+// text file, so we need some parsing functions to parse the content of that
+// text file, and convert then to a `YAML::Node` in our code. A `YAML::Node` is
+// something similar to a `std::map` in C++ or a dictionary in Python. The
+// required parsing functions are in the header `parser.h` and under the namespace
+// `parser`.
+# include <petibm/parser.h>
+
+// The header `mesh.h` defines classes and functions regarding mesh objects under
+// the namespace `mesh`. And it also defines a data type called `Mesh` to unify
+// the member function interface for all different kinds of meshes. This data
+// type, `Mesh`, is under the namespace `type`. In other words, with `type::Mesh`,
+// high-level CFD code developers don't have to worry about what's the underlying
+// mesh. A factory function, `mesh::createMesh`, can create an instance of
+// `type::Mesh` with correct underlying mesh types according to the simulation
+// configurations provided by end-users.
+# include <petibm/mesh.h>
+
+// The header `boundary.h` defines classes and functions regarding boundary
+// conditions and ghost points. A data type called `Boundary` is defined under
+// the namespace `type`. An instance of `type::Boundary` can help developers
+// to call its member functions in a collective way under parallel environments.
+# include <petibm/boundary.h>
+
+// This header defines classes and functions regarding objects holding solutions.
+// A data type called `Solution` is defined under `type` namespace to unify the
+// interface to member functions of different kinds of solution structure.
+# include <petibm/solution.h>
+
+// This header defines functions to create different kinds of operators (i.e.,
+// matrices). In PetIBM, we try to use linear algebra to express calculations.
+// So even the non-linear operators, such as the convection term, are designed
+// as matrix-free operators (matrix-free matrices).
+# include <petibm/operators.h>
+
+// `linsolver.h` defines classes and functions for linear solvers. It alse defines
+// a top-level data type `LinSolver` under the namespace `type` to unify the
+// interface of different kinds of underlying linear solvers. The factory
+// function `createLinSolver` will create the correct underlying linear solver
+// for a `type::LinSolver` instance according to the simulation configurations.
+# include <petibm/linsolver.h>
+
+
+int main(int argc, char **argv)
+{
+    // We will use this variable to hold returned error code for all functions.
+    PetscErrorCode ierr;
+
[…]
