### 2018-05-02 18c6bc6
add README.md to the N-S api example
