### 2018-05-02 1d45b75
add docs regarding using API
