+++ examples/api_examples/README.md
@@ -1,0 +1,14 @@
+# API examples
+
+This folder contains the examples of using PetIBM API to build flow solvers.
+The solvers provided in this folder are just for demonstration purpose. While
+they may still be used to conduct flow simulations, their functionalities are 
+limited.
+
+The solvers with more complete features can be found in the application
+folder under PetIBM source folder. And their executable binary can be found in 
+*bin* folder after building and installing PetIBM.
+
+The basic example is the Navier-Stokes solver in the `liddrivencavity2d`. And
+in `oscillatingcylinder2dRe100_GPU` is the solver solving the flow over
+an oscillating cylinder in 2D and using GPU.
