### 2018-05-02 c1526b0
fix typos in markdown files
