### 2018-05-03 777cdb5
Revise readme and doc for PetIBM API usage
