+++ doc/markdowns/usepetibmapi.md
@@ -1,95 +1,78 @@
 # Use PetIBM API
 
-In the installation path, PetIBM also provides a library and header files that 
-have all necessary API to build a flow solver.
+In the installation path, PetIBM also provides a library and header files that have all necessary API to build a flow solver.
 
 ## API Documentation
 
-The API documentation can be found in the **Modules** section of the 
-[Doxygen pages](https://barbagroup.github.io/PetIBM/modules.html). The subsection
-of **PetIBM building blocks** shows the details of the elements for building
-a flow solver. And the subsection of **Flow solvers and utilities** shows the
-flow solvers built on those building blocks.
+The API documentation can be found in the **Modules** section of the [Doxygen pages](https://barbagroup.github.io/PetIBM/modules.html).
+The subsection of **PetIBM building blocks** shows details of the elements for building a flow solver.
+The subsection of **Flow solvers and utilities** shows the flow solvers built on those building blocks.
 
-A local version of Doxygen pages can be created by `doxygen` utility. Go into
-the **doc** folder of PetIBM, and run `doxygen` command. HTML pages will be
-created and put into a subfolder called **html**. Users then can open up the 
-main page of the manual, **html/index.html**, using a preferred browser. 
+A local version of Doxygen pages can be created with the `doxygen` utility.
+To do that, go into the `doc` folder of PetIBM, and run the `doxygen` command.
+HTML pages will be created in the subfolder `html` of the `doc` directory.
+Users can, then, open the main page of the manual, `html/index.html`, using their preferred browser. 
 
 ## API usage examples
 
-The source code of the flow solvers mentioned in the API documentation may be
-unfriendly for beginners, so we provide other simplified solvers in the folder
-**$PETIBM_SOURCE/examples/api_examples**. 
-The one in the sub-folder **liddrivencavity2d**
-provides step-by-step instruction in the comments of the source code. See the
-README files in those folders.
+The source code of the flow solvers mentioned in the API documentation may be unfriendly for beginners, so we provide other simplified solvers in the folder `$PETIBM_DIR/examples/api_examples`.
+For example, `liddrivencavity2d` provides step-by-step instructions in the comments of the source code.
+(See the README files in those folders for additional details.)
 
 ## Compiling and linking against PetIBM
 
 ### Header files
 
-Let **$PETIBM_INSTALL** be the installation path of PetIBM. All header files for
-public PetIBM API are under **$PETIBM_INSTALL/include/petibm**, while those for
-third-party dependencies, if any, are under **$PETIBM_INSTALL/include**. For example,
-if users use `--enable-amgxwrapper` flag while configuring PetIBM, the 
-`AmgXSolver.hpp` will show up in **$PETIBM_INSTALL/include**.
+Let `$PETIBM_INSTALL` be the installation path of PetIBM.
+All header files for public PetIBM API are under `$PETIBM_INSTALL/include/petibm`, while those for third-party dependencies, if any, are under `$PETIBM_INSTALL/include`.
+For example, if users configure PetIBM using the flag `--enable-amgxwrapper`, the header file `AmgXSolver.hpp` will be copied to `$PETIBM_INSTALL/include`.
 
 There are many headers of PetIBM, while only the top-level ones are normally
 used in an application code:
 
-* `type.h`: useful data types in PetIBM
-* `parser.h`: functions to parse input YAML file
-* `mesh.h`: abstract mesh object and its factory function
[…]
