+++ examples/api_examples/liddrivencavity2d/README.md
@@ -1,68 +1,47 @@
 # Example of using PetIBM API: a simple Navier-Stokes solver
 
-This folder contains a `main.cpp` file which implements a simple Navier-Stokes
-solver with block LU decomposition type of projection method shown in Perot's paper
-(1993). The aim here is to demonstration the basic usage of PetIBM API for those
-who want to use PetIBM as a library or toolbox to build their own flow solvers.
+This folder contains a `main.cpp` file which implements a simple Navier-Stokes solver with block LU decomposition type of projection method shown in Perot's paper (1993).
+The aim here is to show a basic usage of PetIBM API for those who want to use PetIBM as a library or toolbox to build their own flow solvers.
 
-In order to simplify the code in this example, the temporal integration is 
-hard-coded with 1st order implicit and explicit Euler methods for diffusion and
-convection terms, respectively. And some flow and numerical parameters are also
-hard coded, such as viscosity and time-step size. So this example can focus on
-the following concepts:
+In order to simplify the code in this example, the temporal integration is hard-coded with first-order implicit and explicit Euler methods for diffusion and convection terms, respectively.
+Some flow and numerical parameters are also hard-coded, such as the viscosity and the time-step size.
+Thus, this example can focus on the following concepts:
 
 1. Reading and parsing a YAML text file that has simulation parameters in it.
-2. Creating basic objects: mesh, boundary conditions (ghost points), solution,
-   linear solvers.
+2. Creating basic objects: mesh, boundary conditions (ghost points), solution, linear solvers.
 3. Creating operators (matrices).
 
-Following the comments in the `main.cpp`, readers should be able to understand
-the basic usage of PetIBM API. For more details of the classes, functions, and
-namespaces in PetIBM, please refer to 
-[API documentation](https://barbagroup.github.io/PetIBM/modules.html).
+By following the comments in the `main.cpp`, readers should be able to understand the basic usage of PetIBM API.
+For more details about classes, functions, and namespaces in PetIBM, please refer to the [API documentation](https://barbagroup.github.io/PetIBM/modules.html).
 
-This example omits many elements that a complete CFD solver should have, such as
-restarting, outputting transient solutions, flexible temporal integration schemes,
-etc. To see how to implement these elements, please refer to the source code of
-the Navier-Stokes solver in the application folder under the PetIBM source folder.
+This example omits many elements that a feature-complete CFD solver should have, such as restarting, outputting transient solutions, flexible temporal integration schemes, etc.
+To see how to implement these elements, please refer to the source code of the Navier-Stokes solver in the `applications` folder under the PetIBM main directory.
 
 ## Build this example
 
-We provide a shell script, `build.sh` for readers to build this example. It's
-not our intention to demonstration how to compile and build software under Linux
-environment, so we don't provide any Makefile or CMake files in this example.
-But the shell script should be enough to show how to link against PetIBM library.
+We provide a shell script, `build.sh`, for readers to build this example.
+It's not our intention to demonstration how to compile and build software under Linux environment, so we don't provide any Makefile or CMake files in this example.
+But the shell script should be enough to show how to link against the PetIBM library.
 
-In a nutshell, when using PetIBM as a library, we also need to link against PETSc
-library, AmgX library (if used), and other user-provided third-party libraries (
-for example, if users provide their own yaml-cpp, instead of using the one 
-provided by PetIBM configuration system).
+In a nutshell, when using PetIBM as a library, we also need to link against PETSc library, AmgX library (if used), and other user-provided third-party libraries (for example, if users provide their own yaml-cpp, instead of using the one provided by PetIBM configuration system).
 
-In order to use `build.sh`, readers should first modify the variables in the
-first section of the scrip to reflect the path of PETSc and PetIBM installation. 
[…]
