### 2018-05-03 e8becee
Correct small typos in code comments
