### 2018-05-04 #558 @mesnardo
@jedbrown, thank you for those useful comments.
Below is our reply to them.

> I don't know why the 2D simulation is run via petibm-tairacolonius while the 3D uses petibm-decoupledibpm. Do these need to be separate executables and are these the best names for a user?

The executables `petibm-tairacolonius` and `petibm-decoupledibpm` both work with 2D and 3D configurations.
`petibm-tairacolonius` solves the Navier-Stokes equations with an Immersed Boundary Projection Method (IBPM) proposed by Taira and Colonius (2007).
`petibm-decoupledibpm` is a variant of the IBPM, proposed by Li et al. (2016), that decouples the pressure field from the Lagrangian boundary forces.
Currently, there is no 3D example using the IBPM approach (`petibm-tairacolonius`) incorporated in the software package (folder `examples/ibpm`).
In the folder `examples/decoupledibpm`, there are 2D and 3D examples using the decoupled approach (`petibm-decoupledibpm`).

For consistency, we have renamed the executable `petibm-tairacolonius` into `petibm-ibpm` (examples and documentation have also been update  to reflect this change).
We have also added more details about the program executables installed by the package to the documentation (file `runpetIBM.md`).
(There, we now mention explicitly that those programs work for 2D and 3D configurations.)

From the perspective of using PetIBM as a library, the source code of the solvers serve as coding examples.
And separating different methods into different executables makes the source code of each solver more understandable.
When PetIBM's users read the source code and try to learn how to write a solver using PetIBM, it's easier for them to follow the code of a single solver at a time.

_References:_
* Taira, K., & Colonius, T. (2007). The immersed boundary method: a projection approach. Journal of Computational Physics, 225(2), 2118-2137.
* Li, R. Y., Xie, C. M., Huang, W. X., & Xu, C. X. (2016). An efficient immersed boundary projection method for flow over complex/moving boundaries. Computers & Fluids, 140, 122-135.

> It would be a lot more convenient (given vis user interfaces) if there was a single Xdmf file containing all the variables instead of separate files for each variable.

We agree, it would be more convenient.
With PetIBM, the Navier-Stokes equations are solved on staggered grids and we have not (yet) found a working solution to have several multiple `Domain` blocks in the same XDMF file that are readable by VisIt or ParaView.

> The package installs headers and a library (shared and static, plus libtool), suggesting that calling the library (rather than running the provided executables) is an intended mode of operation. However, there is no documentation of this, instead focusing exclusively on the provided executables. If the libraries are not intended to be called by user code, I would recommend not installing the headers and only installing the used library libpetibm.so.0. If it is intended for use, then it should be documented and the library should be cleaned of non-namespaced symbols (e.g., readSingleYAML).

PetIBM is primarily a library (a toolbox to solve the Navier-Stokes equations with an immersed-boundary method) and we use it to develop application codes to implement different types of immersed-boundary methods.
The package installs headers, a library, and executables of the immersed-boundary methods implemented.
From the user perspective, the binary programs can be directly used to compute and post-process the flow using one of the already implemented immersed-boundary methods.
The headers and library can be used to create new application codes.

In the `examples` folder, we have added a sub-folder `api_examples` that contains two examples of application codes using the library: a simple Navier-Stokes solver and an oscillating-cylinder example.
(The Navier-Stokes example contains a README and extensive code comments to guide the user on how to use the API.)
In the documentation, we have added notes regarding the usage of the PetIBM library (file `usepetibmapi.md`).

The non-namespaced symbols are private functions used only by other PetIBM public functions. Their prototypes are not in any headers installed. And they are not shown in the Doxygen pages.
