### 2018-05-04 109ebc6
typo in runpetibm.md
