### 2018-05-04 #558 @jedbrown
Thanks for the explanation and improved documentation. I wonder if there could be more clarity about what needs to be done differently to run ibpm versus decoupledibpm. I see very little substantive differences (solver configuration) but lots of cosmetic differences (comments, subtly different script behavior).

```
cd examples
git diff --no-index ibpm/cylinder2dRe100_GPU/ decoupledibpm/cylinder2dRe100_GPU/
 {ibpm => decoupledibpm}/cylinder2dRe100_GPU/README.md |  8 ++++----
 .../cylinder2dRe100_GPU/config.yaml                   |  3 +++
 .../cylinder2dRe100_GPU/scripts/createBody.py         |  9 +++------
 .../scripts/plotForceCoefficients.py                  | 19 +++++++++----------
 .../cylinder2dRe100_GPU/solversAmgXOptions.info       | 12 ++++++++----
 .../cylinder2dRe100_GPU/solversPetscOptions.info      |  5 +++++
 6 files changed, 32 insertions(+), 24 deletions(-)
```
This organization makes it more difficult to compare methods when it could have been as easy as running the same test case with different solvers. It's also a lot more to maintain and the small inconsistencies grow over time.

My concern with namespacing is link-time collision. For example, the user or a different library could define a symbol with the same name and yet it might end up calling your version. Either make them static, internal in the sense of GCC `__attribute__((visibility("hidden")))`, or namespaced.

What do you think about installing a pkg-config file? That is more standard than relying on the libtool file, and makes linking relatively easy for users that don't use libtool (including CMake).
