### 2018-05-07 d4ff49d
use anonymous namespace to apply internal linkages
