+++ src/parser/parser.cpp
@@ -14,14 +14,111 @@
 # include <petibm/yaml.h>
 # include <petibm/misc.h>
 
-// private function. Read config.yaml and other files for overwriting
-PetscErrorCode readYAMLs(YAML::Node &node);
+
+namespace // anonymous namespace for internal linkage
+{
 
 // private function. Read a single YAML file overwriting part of config.yaml
-PetscErrorCode readSingleYAML(YAML::Node &node, const std::string &s);
+PetscErrorCode readSingleYAML(YAML::Node &node, const std::string &s)
+{
+    PetscFunctionBeginUser;
+
+    YAML::Node  temp;
+
+    if (node[s+".yaml"].IsDefined())
+    {
+        try
+        {
+            temp = YAML::LoadFile(node[s+".yaml"].as<std::string>());
+        }
+        catch(YAML::BadFile &err)
+        {
+            SETERRQ2(PETSC_COMM_WORLD, PETSC_ERR_FILE_OPEN,
+                "Unable to open %s"
+                "when reading settings from %s.yaml\n",
+                node[s+".yaml"].as<std::string>().c_str(), s.c_str());
+        }
+
+        if (! node[s].IsDefined()) node[s] = YAML::Node();
+        for(auto it: temp) node[s][it.first] = it.second;
+    }
+
+    PetscFunctionReturn(0);
+} // readSingleYAML
+
+
+// private function. Read config.yaml and other files for overwriting
+PetscErrorCode readYAMLs(YAML::Node &node)
+{
+    PetscFunctionBeginUser;
+
+    PetscErrorCode ierr;
+    YAML::Node temp;
+
+    // open config.yaml. If failed, return error through PETSc and terminate.
+    try
+    {
+        temp = YAML::LoadFile(node["config.yaml"].as<std::string>());
+    }
+    catch(YAML::BadFile &err)
+    {
+        SETERRQ1(PETSC_COMM_WORLD, PETSC_ERR_FILE_OPEN,
+            "Unable to open %s "
+            "when reading settings from config.yaml\n",
+            node["config.yaml"].as<std::string>().c_str());
+    }
+
+    // copy what we read from config.yaml to the `node`
+    for(auto it: temp) node[it.first] = it.second;
+
+    // overwrite mesh section if mesh.yaml is provided
+    ierr = readSingleYAML(node, "mesh"); CHKERRQ(ierr);
+
+    // overwrite flow section if flow.yaml is provided
+    ierr = readSingleYAML(node, "flow"); CHKERRQ(ierr);
+
+    // overwrite parameters section if parameters.yaml is provided
+    ierr = readSingleYAML(node, "parameters"); CHKERRQ(ierr);
+
+    // overwrite bodies section if bodies.yaml is provided
+    ierr = readSingleYAML(node, "bodies"); CHKERRQ(ierr);
+
+    PetscFunctionReturn(0);
+} // readYAMLs
+
 
 // private function. Check if a directory exists. Create one if not.
-PetscErrorCode checkAndCreateFolder(const std::string &dir);
+PetscErrorCode checkAndCreateFolder(const std::string &dir)
+{
+    PetscFunctionBeginUser;
+
+    PetscErrorCode      ierr;
+    PetscBool           flg;
+
+    ierr = PetscTestDirectory(dir.c_str(), 'w', &flg); CHKERRQ(ierr);
+
+    if (! flg)
+    {
+        PetscMPIInt     rank;
+        ierr = MPI_Comm_rank(PETSC_COMM_WORLD, &rank); CHKERRQ(ierr);
+
+        if (rank == 0)
+        {
+            int status;
[…]

[…]
