### 2018-05-08 #558 @labarba
The substantive difference between running a case with ibpm versus decoupledibpm lies in the mathematical formulation, which results in different linear systems (even different number of systems) to solve. This difference reflects in the choice of solver and preconditioner that, in each case, will perform better, given the structure of the matrices. At this point, a user needs to read some papers to understand something about each method.

Since PetIBM is a research tool, users need to do this extra work of studying the methods, together with running the examples we provide. The two method require different solver configuration (explained in the README).

We keep an example folder per executable. This keeps each method self-contained in the repo. (A user may want to work with only one of the methods.) I guess you're saying that it might be nice to have a common set of examples for both methods. It didn't occur to us, and we have no strong preference, but at this point, it feels like unnecessary work to reorganize the repo.
