### 2018-05-08 #558 @labarba
Creating a pkg-config script seems like overkill for this project. Bear in mind, at this point we have no known users outside the group. I suppose if some user appeared who prefers this, we'd ask them to contribute it via PR!
