### 2018-05-08 #558 @piyueh
Regarding the functions not in any namespace (i.e., local functions), we fixed this in commit [e38da005](https://github.com/barbagroup/PetIBM/commit/e38da005e0a7a91bf9128baad3f491a5ce7cec1e). We use anonymous namespace to encapsulate these local functions and force internal linkage only, so API users will not be able to use these local functions.
