### 2018-05-08 #558 @kyleniemeyer
@labarba @jedbrown regarding the pkg-config point, I would argue that one of the purposes of the review here is to make recommendations to make software more user-friendly outside the immediate developers/group, so it seems like a useful suggestion to consider from that point of view.
