### 2018-05-08 #558 @labarba
@kyleniemeyer Reasonable point, in general, but libtools is a perfectly acceptable solution here. No one in our group has gone through the exercise of learning pkg-config, so it would be starting from scratch with that tool.

Users of PetIBM need to also link to PETSc, which provides custom scripts for resolving its dependencies, and the user will need to learn about that anyway. PetIBM has almost the same dependencies as PETSc (only real difference is AmgX).
