### 2018-05-08 #558 @jedbrown
PETSc provides a pkg-config. It isn't a difficult standard and is straightforward to provide from autotools: https://autotools.io/pkgconfig/file-format.html

@piyueh Thanks for that fix. I think these three symbols are still not intended to be exported in a non-petibm namespace:

```
000000000004c4cd T operator<(MatStencil const&, MatStencil const&)
0000000000068c76 T YAML::convert<PetscBool>::decode(YAML::Node const&, PetscBool&)
0000000000068dfc T YAML::convert<PetscBool>::encode(PetscBool const&)
```

@labarba On organization of examples, this is certainly not a blocking issue, but I felt it was worth pointing out the maintenance burden (leading to the observed divergence) of separate examples when only a few solver parameters are actually different. I think it would be better for both developers and users to unify everything that can be, but I'm happy to accept the JOSS paper either way.
