### 2018-05-08 #558 @labarba
PETSc does provide pkg-config, but PETSc is a 20+ year old project, with thousands of users around the world. From what we've seen the PETSc trainings and documentation explain how to use the customized makefiles (we use that). The manual mentions only in passing that pkg-config exists (section 15.13 Qt Creator Users).

It may not be a _difficult_ standard, but what is the ROI of adding that now? Is it just "nice to have"?
