### 2018-05-08 #558 @jedbrown
Pkg-config immediately works with CMake, plain makefiles, and any other build system. The libtool files are mainly usable only by autotools. So providing pkg-config significantly reduces the effort and improves reliability of linking for anyone who is not using autotools.
