### 2018-05-08 #558 @labarba
You assert that pkg-config has advantages over libtool, and I hear you. It still sounds as "nice to have" and not a necessity.  Given that this submission is 3+ months in the pipeline (high for JOSS), considering also that a majority of accepted JOSS papers report software projects of smaller scale, and submitting that the functionality is already provided (only with a "lesser" tool), I respectfully propose that you don't require this for acceptance.

I propose we open an issue on the repo, but not allocate researcher time right now for this improvement.

As justification, I propose that users of PetIBM do require a higher level of skill than the majority of users of JOSS-published software: it relies on PETSc, it uses GPUs, it solves 3D Navier-Stokes ... We are comfortable with the effort that is required for someone who wants to try PetIBM given the ecosystem they inhabit.

We also plan to continue working on the project. Currently, it has 32 stars on GitHub… we may get some comments on an open issue, if someone else is keen on the improvement.

But for now, I request that you let us publish without it.
