### 2018-05-08 efc110d
make YAML converters and operator< of MatStencil private
