### 2018-05-08 9e701d4
update configure and Makefile.in
