### 2018-05-08 #558 @piyueh
@jedbrown Thank you for pointing out those three functions! These three functions should now be fixed in the current state of `joss-review` (from commit [9e701d](https://github.com/barbagroup/PetIBM/commit/efc110d834074ef510627cb99e3231afc5b1f7a4)).

Actually, we decided to put all YAML converters to private functions, not just the two for `PetscBool` mentioned. They are not for public API users by current design.
