### 2018-05-08 #558 @kyleniemeyer
It looks like the pkg-config option is something that would be nice for the future, but it seems is not required to build and use the software now. Do you agree @jedbrown? If so, have your other issues been resolved?
