### 2018-05-09 #558 @jedbrown
Yeah, fine by me.On May 8, 2018 17:14, Kyle Niemeyer <notifications@github.com> wrote:It looks like the pkg-config option is something that would be nice for the future, but it seems is not required to build and use the software now. Do you agree @jedbrown? If so, have your other issues been resolved?

—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or mute the thread.
