### 2018-05-09 #558 @kyleniemeyer
@labarba @mesnardo @piyueh a few comments on the paper before we accept:

- Can you complete your affiliations? (i.e., add city, state, country)
- Generally the first paragraph is a bit long, and can probably be broken into multiple paragraphs which are structured a bit better.
- You do not actually define IBM or GPU prior to their first use.
- The comment about pressure boundary conditions having "caused many headaches for CFD practitioners!" seems a bit casual/opinionated, and isn't backed up by anything. I would suggest either removing that bit, or explaining/backing it up further (which may not be appropriate in this sort of paper).
- The sentence "PetIBM features an operator-based design, so it can be used as a toolbox for researching new solution methods." is not clear to me—can you elaborate a bit more on this?
- I think there are a few details mentioned in the README that do not appear in the paper, such as support for multiple immersed bodies and moving bodies with prescribed kinetics; it may help to discuss such support. It also might be nice to mention the strong set of examples included.
