### 2018-05-09 def57f8
complete the affiliation (department, city, state, country)
