### 2018-05-12 c36726e
Update with AmgXWrapper-1.4
* Update version of AmgXWrapper (from `1.0-beta2` to `1.4`).
* Update calls to AmgXWrapper routines (due to changes in API).
* AmgXWrapper-1.4 requires to use PETSc-3.8+. Therefore, when configuring with AmgX, use PETSc-3.8+.
* When checking for AmgX, find and keep only the first occurrence of `libamgxsh.so`.
* When configuring with AmgX, do not check the version of OpenMPI anymore.
* Do not check for Doxygen.
* Bump version of PetIBM to 0.2.1
