+++ external/AmgXWrapper-1.4/paper.bib
@@ -1,0 +1,21 @@
+@misc{chuang+barba2017,
+    doi = {10.6084/m9.figshare.5018774.v1},
+    author = {Chuang,  Pi-Yueh and Barba,  Lorena A.},
+    keywords = {Computational Fluid Dynamics, Numerical Computation, Distributed Computing},
+    title = {Using AmgX to Accelerate PETSc-Based CFD Codes},
+    publisher = {Figshare},
+    year = {2017}
+}
+
+@TechReport{petsc-user-ref,
+    author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Jed Brown and Peter Brune
+                and Kris Buschelman and Lisandro Dalcin and Victor Eijkhout and William~D. Gropp
+                and Dinesh Kaushik and Matthew~G. Knepley
+                and Lois Curfman McInnes and Karl Rupp and Barry~F. Smith
+                and Stefano Zampini and Hong Zhang and Hong Zhang},
+    title  = {{PETS}c Users Manual},
+    institution = {Argonne National Laboratory},
+    year   = 2016,
+    number = {ANL-95/11 - Revision 3.7},
+    url    = {http://www.mcs.anl.gov/petsc}
+}
