### 2018-05-18 91b8fe8
Revise JOSS paper
based on suggestions from @kyleniemeyer in openjournals/joss-reviews#558.

* Split first paragraph into three (general info about PetIBM, IBMs implemented, and comparison with other open-source software).
* Define IBM and GPU.
* Add references to the work of Gresho and Sani regarding pressure boundary conditions for incompressible flows.
* Elaborate on the operator-based implementation.
* Add details about what can be done with PetIBM (multiple, moving, etc.).
* Add note about user's documentation and presence of many examples.
