### 2018-05-23 64231d1
Fix bug with MPI barrier when creating directory
* Do not call `PetscTestDirectory`.
* Have all processes trying to create the dir (with `mkdir`).
* Call error handler if dir not created (for reason other than "File exists").
* Rename private function: `checkAndCreateFolder` -> `createDirectory`.
