### 2018-05-24 1557286
Add equations in appendix; remove extra comma
