+++ paper.md
@@ -47,9 +47,346 @@ A previous project implementing immersed-boundary methods on GPU architecture is
 
 PetIBM is written in C++ and relies on the PETSc library (@petsc_1997, @petsc_user_ref_2017) for data structures and parallel routines to run on memory-distributed architectures.
 PetIBM can solve one or several linear systems on multiple distributed CUDA-capable GPU devices with the NVIDIA AmgX library and AmgXWrapper (@chuang_barba_2017).
-The software package includes extended documentation as well as, many examples to guide users.
+The software package includes extended documentation as well as many examples to guide users.
 
 PetIBM has already been used to generate results published in @mesnard_barba_2017, a full replication of a study on the aerodynamics of a gliding snake species (@krishnan_et_al_2014).
 PetIBM is currently used to compute the three-dimensional flow of a gliding-snake model on the cloud platform Microsoft Azure.
+
+## Appendix: mathematical formulation
+
+PetIBM solves the Navier-Stokes equations on an extended discretization grid that includes the interior of the immersed boundary.
+To model the presence of the boundary, a forcing term is added to the momentum equation and an additional equation for the no-slip condition completes the system.
+Many variants of the immersed-boundary method (IBM) depend on one models the forcing.
+In PetIBM, we use regularized-delta functions to transfer data between the Eulerian grid and the Lagrangian boundary points.
+The system of equation is:
+
+\begin{equation}
+\begin{cases}
+\frac{\partial \mathbf{u}}{\partial t} + \mathbf{u} \cdot \nabla \mathbf{u} = -\nabla p + \frac{1}{Re} \nabla^2 \mathbf{u} + \int_{s}{\mathbf{f} \left( \mathbf{\xi} \left( \mathit{s}, \mathit{t} \right) \right) \delta \left( \mathbf{\xi} - \mathbf{x} \right)} d\mathit{s} \\
+\nabla \cdot \mathbf{u} = 0 \\
+\mathbf{u} \left( \mathbf{\xi} \left( \mathit{s}, t \right) \right) = \int_{\mathbf{x}}{\mathbf{u} \left( \mathbf{x} \right)} \delta \left( \mathbf{x} - \mathbf{\xi} \right) d\mathbf{x}
+\end{cases}
+\end{equation}
+
+where $\mathbf{u}$ is the velocity field, $p$ is the fluid pressure, and $Re$ is the Reynolds number.
+
+Currently, PetIBM provides two application codes implementing different versions of the IBM: (1) an immersed-boundary projection method (IBPM) based on the work of @taira_colonius_2007 and (2) a decoupled version of the IBPM proposed by @li_et_al_2016.
+Those two methods fit into the framework of the projection approach of @perot_1993.
+The equations are fully discretized (space and time) to form an algebraic system to be solved for the velocity $u^{n+1}$, the pressure field $\phi$, and the Lagrangian forces $\tilde{f}$.
+The discretized system is:
+
+\begin{equation}
+\left[
+\begin{matrix}
[…]
