### 2018-05-24 #558 @labarba
@kyleniemeyer We're done with our beautiful paper!
