### 2018-05-25 #558 @kyleniemeyer
Wow! Big changes. It is just about there, just

In the appendix, this sentence needs to be fixed (maybe missing "how"?):
> Many variants of the immersed-boundary method (IBM) depend on one models the forcing.

and then "system of equation" -> "system of equations" two sentences later, and "as follow" -> "as follows" before equation 4.

With those minor fixes, we'll be good to go.
