### 2018-05-25 3b1fe76
Fix typos in JOSS paper
