### 2018-05-25 242aee6
Update changelog; bump version; run autogen
Bump version to 0.3.1.
