### 2018-05-29 #558 @mesnardo
@kyleniemeyer PetIBM version bumped to 0.3.1 and archived on Zenodo (10.5281/zenodo.1255132).
